+++ README.md
@@ -22,3 +22,4 @@ Dica: Utilize o catálogo para pensar em possíveis sugestões de refatoração 
 - Remoção dos números mágicos
 - Extração da superclasse
 - Encapsulamento dos atributos das classes
+- Modularização
