--- include/Cobertura.hpp
@@ -1,19 +1,0 @@
-#ifndef COBERTURA_HPP
-#define COBERTURA_HPP
-
-#include "Imovel.hpp"
-#include "Cliente.hpp"
-using namespace std;
-
-class Cobertura : public Imovel {
-    
-    public:
-        Cobertura(string corretor, double area, int quartos, int banheiros, int vagas, 
-                  double  valor_por_metro_quadrado, Cliente vendedor);
-        void print();
-
-    private:
-        const double kTaxaComissao = 0.10;
-};
-
-#endif
